# cmd/docs/swagger.yaml
basePath: /api
definitions:
...
  handler.AddMaterialToCartRequest:
    properties:
      material_id:
        type: integer
    required:
    - material_id
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Бэкенд сервис для анализа материалов и радиоуглеродного датирования.
  title: Система анализа материалов
  version: "1.0"
paths:
...
  /dating/cart:
    get:
      description: Get current user's dating cart with materials
      produces:
      - application/json
      responses:
        "200":
          description: Success response with cart data
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get dating cart
      tags:
      - Dating Cart
  /dating/cart/icon:
    get:
      description: Get the number of materials in dating cart
      produces:
      - application/json
      responses:
        "200":
          description: Success response with count
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get cart item count
      tags:
      - Dating Cart
  /dating/cart/materials:
    post:
      consumes:
      - application/json
      description: Add material to dating cart
      parameters:
      - description: Material data
        in: body
        name: material
        required: true
        schema:
          $ref: '#/definitions/handler.AddMaterialToCartRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Material added successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Material not found
          schema:
            additionalProperties: true
            type: object
        "409":
          description: Material already in cart
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Add material to cart
      tags:
      - Dating Cart
...
schemes:
- http
securityDefinitions:
  BearerAuth:
    description: 'Используется для запросов через Insomnia/Postman: "Bearer <JWT>"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"